Fix issue with intervals above 2**31 (#38)

+++ ncls/src/cncls.pxd
@@ -52,8 +52,8 @@ cdef extern from "ncls/src/intervaldb.h":
         int start
         int len
 
-    int find_overlap_start(int64_t start, int64_t end, IntervalMap im[], int n)
-    int find_suboverlap_start(int64_t start, int64_t end, int isub, IntervalMap im[], SublistHeader subheader[])
+    int64_t find_overlap_start(int64_t start, int64_t end, IntervalMap im[], int n)
+    int64_t find_suboverlap_start(int64_t start, int64_t end, int isub, IntervalMap im[], SublistHeader subheader[])
     int imstart_qsort_cmp(void *void_a,void *void_b)
     IntervalMap *read_intervals(int n,FILE *ifile)
     SublistHeader *build_nested_list(IntervalMap im[],int n,int *p_n,int *p_nlists)
